oauth/src/main/java/com/knowledgebase/oauth/user/UserService.java
package com.knowledgebase.oauth.user;

import java.util.Arrays;
import java.util.HashSet;
import java.util.Optional;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
import org.springframework.stereotype.Service;
import com.knowledgebase.oauth.role.Role;
import com.knowledgebase.oauth.role.RoleName;
import com.knowledgebase.oauth.role.RoleRepository;

@Service("userService")
public class UserService {

	private UserRepository userRepository;
	private RoleRepository roleRepository;
	private BCryptPasswordEncoder bCryptPasswordEncoder;

	@Autowired
	public UserService(UserRepository userRepository, RoleRepository roleRepository,
			BCryptPasswordEncoder bCryptPasswordEncoder) {
		this.userRepository = userRepository;
		this.roleRepository = roleRepository;
		this.bCryptPasswordEncoder = bCryptPasswordEncoder;
	}

	public Optional<User> findUserByEmail(String email) {
		return userRepository.findByEmail(email);
	}

	public User saveUser(User user) {
		user.setPassword(bCryptPasswordEncoder.encode(user.getPassword()));
		user.setActive(1);
		Optional<Role> userRole = roleRepository.findByName(RoleName.ROLE_ADMIN);
		user.setRoles(new HashSet<>(Arrays.asList(userRole.get())));
		return userRepository.save(user);
	}

}
